python-srv/requirements.txt
Flask==2.3.*
hdbcli==2.23.*
generative-ai-hub-sdk[all]
ai_core_sdk
hana-ml==2.23.*
python-dotenv==1.0.*
gunicorn
cfenv
sap-xssec